Feat: Add structured properties LUT and monitoring integration
- Add Option B structured properties to TernarySpace (digit_count, parent, level_rank, etc.)
- Add HardwareMonitor class for GPU/RAM tracking with OOM diagnostics
- Integrate TensorBoardLogger and progress bars into train.py
- Fix config wiring (option_c, StateNet params, device)
- Enforce float64 precision throughout for geoopt compatibility
- Update terminology: freeze/unfreeze → fix/train, frozen_checkpoint → anchor_checkpoint
- Clean up 5.12.4.yaml, remove dead config sections
- Add requirements.txt with core dependencies
- Update documentation with new features

+++ src/geometry/poincare.py
@@ -231,7 +231,7 @@ def geodesic_interpolation(x: torch.Tensor, y: torch.Tensor, steps: int = 10, c:
         Points along geodesic, shape (steps, ..., dim)
     """
     manifold = get_manifold(c, device=x.device)
-    t_values = torch.linspace(0, 1, steps, device=x.device)
+    t_values = torch.linspace(0, 1, steps, device=x.device, dtype=x.dtype)
     return torch.stack([manifold.geodesic(t.item(), x, y) for t in t_values])
